#lsd
An iPhone app to felicitate the annual freshman tradition called Larkin
Silent Dance (commonly abbreviated as l.s.d) Larkin Silent Dance is
associated with a freshman dorm called Larkin and the idea behind the
event is that every single resident of house downloads a common list of
songs. At the count of 3 usually initiated by an RA, Everyone hits the
Play button and walks around the campus dancing while listening to the
same music on their headphones.